LayoutTests/platform/wpe/storage/domstorage/localstorage/enumerate-with-length-and-key-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
This test attempts to enumerate all the keys in localStorage with .length + .key(). The built-in properties of the Storage object should be ignored. The test operates on the localStorage object.
